Move multisig deploy to withdraw
